+++ ext/GraphNeuralNetworksCUDAExt/GNNGraphs/query.jl
@@ -1,0 +1,2 @@
+
+GNNGraphs._rand_dense_vector(A::CUMAT_T) = CUDA.randn(size(A, 1))
