+++ ext/GraphNeuralNetworksCUDAExt/GNNGraphs/transform.jl
@@ -1,0 +1,2 @@
+
+GNNGraphs.dense_zeros_like(a::CUMAT_T, T::Type, sz = size(a)) = CUDA.zeros(T, sz)
