+++ src/GNNGraphs/GNNGraphs.jl
@@ -2,7 +2,6 @@ module GNNGraphs
 
 using SparseArrays
 using Functors: @functor
-using CUDA
 import Graphs
 using Graphs: AbstractGraph, outneighbors, inneighbors, adjacency_matrix, degree,
               has_self_loops, is_directed
